% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fcn_supported}
\alias{fcn_supported}
\title{Which branches of a network can be run fully convolutionally, exactly.}
\usage{
fcn_supported(model)
}
\arguments{
\item{model}{A dual_branch_cnn.}
}
\value{
A logical per branch.
}
\description{
Which branches of a network can be run fully convolutionally, exactly.
}
