% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoa.R
\name{inside_aoa}
\alias{inside_aoa}
\title{Is each row inside the area of applicability?}
\usage{
inside_aoa(di, threshold)
}
\arguments{
\item{di}{From dissimilarity_index().}

\item{threshold}{From aoa_threshold().}
}
\value{
Logical vector. TRUE means the cross-validated error applies here.
}
\description{
Is each row inside the area of applicability?
}
