% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_bootstrap.R
\name{plot.spatial_correlogram}
\alias{plot.spatial_correlogram}
\title{Plot a spatial correlogram.}
\usage{
\method{plot}{spatial_correlogram}(x, ...)
}
\arguments{
\item{x}{A \code{spatial_correlogram}, from \code{\link[=spatial_correlogram]{spatial_correlogram()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Plot a spatial correlogram.
}
