% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_bootstrap.R
\name{print.block_bootstrap}
\alias{print.block_bootstrap}
\title{Print a \code{block_bootstrap}}
\usage{
\method{print}{block_bootstrap}(x, ...)
}
\arguments{
\item{x}{A \code{block_bootstrap}, from \code{\link[=block_bootstrap]{block_bootstrap()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{block_bootstrap}
}
\keyword{internal}
