% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal.R
\name{print.conformal_cal}
\alias{print.conformal_cal}
\title{Print a \code{conformal_cal}}
\usage{
\method{print}{conformal_cal}(x, ...)
}
\arguments{
\item{x}{A \code{conformal_cal}, from \code{\link[=conformal_calibrate]{conformal_calibrate()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{conformal_cal}
}
\keyword{internal}
