% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal.R
\name{print.conformal_scale}
\alias{print.conformal_scale}
\title{Print a \code{conformal_scale}}
\usage{
\method{print}{conformal_scale}(x, ...)
}
\arguments{
\item{x}{A \code{conformal_scale}, from \code{\link[=conformal_scale_fit]{conformal_scale_fit()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{conformal_scale}
}
\keyword{internal}
