% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal.R
\name{print.conformal_scaled}
\alias{print.conformal_scaled}
\title{Print a \code{conformal_scaled}}
\usage{
\method{print}{conformal_scaled}(x, ...)
}
\arguments{
\item{x}{A \code{conformal_scaled}, from \code{\link[=conformal_scaled_calibrate]{conformal_scaled_calibrate()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{conformal_scaled}
}
\keyword{internal}
