% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal.R
\name{print.cv_plus_cal}
\alias{print.cv_plus_cal}
\title{Print a \code{cv_plus_cal}}
\usage{
\method{print}{cv_plus_cal}(x, ...)
}
\arguments{
\item{x}{A \code{cv_plus_cal}, from \code{\link[=cv_plus_calibrate]{cv_plus_calibrate()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{cv_plus_cal}
}
\keyword{internal}
