% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoa.R
\name{print.di_reference}
\alias{print.di_reference}
\title{Print a \code{di_reference}}
\usage{
\method{print}{di_reference}(x, ...)
}
\arguments{
\item{x}{A \code{di_reference}, from \code{\link[=di_reference]{di_reference()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{di_reference}
}
\keyword{internal}
