% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.dsm_data}
\alias{print.dsm_data}
\title{Print a \code{dsm_data}}
\usage{
\method{print}{dsm_data}(x, ...)
}
\arguments{
\item{x}{A \code{dsm_data}, from \code{\link[=dsm_load]{dsm_load()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{dsm_data}
}
\keyword{internal}
