% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{print.dsm_final}
\alias{print.dsm_final}
\title{Print a \code{dsm_final}}
\usage{
\method{print}{dsm_final}(x, ...)
}
\arguments{
\item{x}{A \code{dsm_final}, from \code{\link[=dsm_final]{dsm_final()}} or \code{\link[=dsm_report_final]{dsm_report_final()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{dsm_final}
}
\keyword{internal}
