% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.dsm_fit}
\alias{print.dsm_fit}
\title{Print a \code{dsm_fit}}
\usage{
\method{print}{dsm_fit}(x, ...)
}
\arguments{
\item{x}{A \code{dsm_fit}, from \code{\link[=dsm_train]{dsm_train()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{dsm_fit}
}
\keyword{internal}
