% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{print.dsm_importance}
\alias{print.dsm_importance}
\title{Print a variable importance.}
\usage{
\method{print}{dsm_importance}(x, n = 20L, ...)
}
\arguments{
\item{x}{A \code{dsm_importance}, from \code{\link[=dsm_importance]{dsm_importance()}}.}

\item{n}{How many variables to show.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a variable importance.
}
\keyword{internal}
