% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{print.dsm_prediction}
\alias{print.dsm_prediction}
\title{Print a \code{dsm_prediction}}
\usage{
\method{print}{dsm_prediction}(x, ...)
}
\arguments{
\item{x}{A \code{dsm_prediction}, from \code{\link[=dsm_predict]{dsm_predict()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{dsm_prediction}
}
\keyword{internal}
