% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{print.dsm_store}
\alias{print.dsm_store}
\title{Print a \code{dsm_store}}
\usage{
\method{print}{dsm_store}(x, ...)
}
\arguments{
\item{x}{A \code{dsm_store}, from \code{\link[=dsm_prepare]{dsm_prepare()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{dsm_store}
}
\keyword{internal}
