% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{print.importance_comparison}
\alias{print.importance_comparison}
\title{Print a comparison of importances.}
\usage{
\method{print}{importance_comparison}(x, ...)
}
\arguments{
\item{x}{An \code{importance_comparison}, from \code{\link[=compare_importance]{compare_importance()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a comparison of importances.
}
\keyword{internal}
