% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{print.importance_spec}
\alias{print.importance_spec}
\title{Print an importance method.}
\usage{
\method{print}{importance_spec}(x, ...)
}
\arguments{
\item{x}{An \code{importance_spec}, from \code{\link[=permutation_importance]{permutation_importance()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an importance method.
}
\keyword{internal}
