% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{print.paired_comparison}
\alias{print.paired_comparison}
\title{Print a \code{paired_comparison}}
\usage{
\method{print}{paired_comparison}(x, ...)
}
\arguments{
\item{x}{A \code{paired_comparison}, from \code{\link[=paired_family_test]{paired_family_test()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{paired_comparison}
}
\keyword{internal}
