% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal.R
\name{print.picp_report}
\alias{print.picp_report}
\title{Print a \code{picp_report}}
\usage{
\method{print}{picp_report}(x, ...)
}
\arguments{
\item{x}{A \code{picp_report}, from \code{\link[=picp_report]{picp_report()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{picp_report}
}
\keyword{internal}
