% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.resample_spec}
\alias{print.resample_spec}
\title{Print a \code{resample_spec}}
\usage{
\method{print}{resample_spec}(x, ...)
}
\arguments{
\item{x}{A \code{resample_spec}, from \code{\link[=spatial_cv]{spatial_cv()}}, \code{\link[=knndm_cv]{knndm_cv()}}, \code{\link[=random_cv]{random_cv()}},
\code{\link[=holdout_cv]{holdout_cv()}} or \code{\link[=region_cv]{region_cv()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{resample_spec}
}
\keyword{internal}
