% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smearing.R
\name{print.smearing_cal}
\alias{print.smearing_cal}
\title{Print a \code{smearing_cal}}
\usage{
\method{print}{smearing_cal}(x, ...)
}
\arguments{
\item{x}{A \code{smearing_cal}, from \code{\link[=smearing_factor]{smearing_factor()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{smearing_cal}
}
\keyword{internal}
