% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smearing.R
\name{print.smearing_check}
\alias{print.smearing_check}
\title{Print a \code{smearing_check}}
\usage{
\method{print}{smearing_check}(x, ...)
}
\arguments{
\item{x}{A \code{smearing_check}, from \code{\link[=smearing_check]{smearing_check()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{smearing_check}
}
\keyword{internal}
