% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{get_model}
\alias{get_model}
\title{Fetch a registered model.}
\usage{
get_model(name)
}
\arguments{
\item{name}{Registered name.}
}
\value{
The \code{model_spec} registered under \code{name}.
}
\description{
Fetch a registered model.
}
\examples{
get_model("rf")
}
