% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_torch_device}
\alias{setup_torch_device}
\title{Configure torch threads and select compute device.}
\usage{
setup_torch_device(n_threads = NULL, use_cuda = TRUE)
}
\arguments{
\item{n_threads}{Number of intra-op threads; NULL for the physical cores
minus one (see resolve_cores()).}

\item{use_cuda}{Use GPU if available.}
}
\value{
A torch_device object.
}
\description{
Configure torch threads and select compute device.
}
